Only do one powf call when computing pitch bend and vibrato modifiers.
Powf is really expensive (when we call it 44100 times per second),
so minimizing the number of times we call it is important. Here, we (and
by we i mean cassie) perform what is known as "algebra skip" to avoid three
extra powf calls.

Specifically, the algebric manipulation is this:
```
  (2 ^ (semis / 12.0)) ^ bend_amount
= 2 ^ (bend_amount * semis / 12.0)
```

Since we are using this function twice for the pitch bend and vibrato
amounts, we actually have:
```
  2 ^ (pitch_bend * max_pitch_bend / 12.0) * 2 ^ (vibrato_env * (vibrato_mod * 2.0) / 12.0)
= 2 ^ ((pitch_bend * max_pitch_bend + vibrato_env * (vibrato_mod * 2.0)) / 12.0)
```

which is only one exponentiation instead of the four we were doing. This
saves nearly a whole second off the megalovania benchmark.